refactor: centralize security headers
