+++ README.md
@@ -124,7 +124,7 @@ docker-compose up
 
 ## 🔒 Security
 
-The application sends several security headers defined in `next.config.ts`:
+The application sends several security headers via middleware (`src/middleware/securityHeaders.ts`):
 
   - **Content-Security-Policy** restricts asset loading to the same origin while
   allowing inline scripts required by Next.js.
