+++ next.config.ts
@@ -1,36 +1,9 @@
 import type { NextConfig } from "next";
 
-const securityHeaders = [
-  {
-    key: 'Content-Security-Policy',
-    value: [
-      "default-src 'self'",
-      "script-src 'self' 'unsafe-inline'",
-      "style-src 'self' 'unsafe-inline' https://fonts.googleapis.com",
-      "img-src 'self' https: data:",
-      "connect-src 'self' https: ws:",
-      "font-src 'self' https://fonts.gstatic.com data:",
-      "frame-ancestors 'none'",
-    ].join('; '),
-  },
-  { key: 'X-Frame-Options', value: 'DENY' },
-  { key: 'X-Content-Type-Options', value: 'nosniff' },
-  { key: 'Referrer-Policy', value: 'same-origin' },
-  { key: 'X-XSS-Protection', value: '1; mode=block' },
-];
-
 const nextConfig: NextConfig = {
   reactStrictMode: true,
   images: {
     domains: ['coin-images.coingecko.com'],
-  },
-  async headers() {
-    return [
-      {
-        source: '/(.*)',
-        headers: securityHeaders,
-      },
-    ];
   },
   /* config options here */
 };
