+++ src/middleware/securityHeaders.ts
@@ -1,0 +1,23 @@
+import { NextResponse } from 'next/server'
+
+export function setSecurityHeaders(res: NextResponse) {
+  const csp = [
+    "default-src 'self'",
+    "script-src 'self' 'unsafe-inline' https:",
+    "style-src 'self' 'unsafe-inline' https://fonts.googleapis.com",
+    "img-src * blob: data:",
+    "connect-src 'self' https:",    
+    "font-src 'self' https://fonts.gstatic.com data:",
+    "object-src 'none'",
+    "frame-ancestors 'none'",
+  ].join('; ')
+
+  res.headers.set('Content-Security-Policy', csp)
+  res.headers.set('Strict-Transport-Security', 'max-age=31536000; includeSubDomains; preload')
+  res.headers.set('Referrer-Policy', 'same-origin')
+  res.headers.set('X-Frame-Options', 'DENY')
+  res.headers.set('X-Content-Type-Options', 'nosniff')
+  res.headers.set('X-XSS-Protection', '1; mode=block')
+  res.headers.set('Permissions-Policy', 'camera=(), microphone=(), geolocation=()')
+  return res
+}
